Replaced make_tests.sh with do_tests.sh.
make_tests.sh created Makefile rules to compile and run the available
tests. It would do this on every invocation of make, and a "make test"
run would stop on the first error.

do_tests.sh will only run on "make test" and try to compile and run all
tests, even if one of them fails. It then returns the total number of
errors encountered. If this is 0, all is well. If it isn't, make will
fail with an error return of its own.